Rename admin search indexListeners and providers
