+++ packages/article/src/Infrastructure/Symfony/HttpKernel/SuluArticleBundle.php
@@ -425,8 +425,8 @@ final class SuluArticleBundle extends AbstractBundle
             ])
             ->tag('sulu_route.route_defaults_provider', ['resource_key' => 'articles']);
 
-        $services->set('sulu_article.article_index_listener')
-            ->class('Sulu\Article\Infrastructure\Sulu\Search\ArticleIndexListener')
+        $services->set('sulu_article.admin_article_index_listener')
+            ->class('Sulu\Article\Infrastructure\Sulu\Search\AdminArticleIndexListener')
             ->args([
                 new Reference('sulu_message_bus'),
             ])
@@ -439,8 +439,8 @@ final class SuluArticleBundle extends AbstractBundle
             ->tag('kernel.event_listener', ['event' => ArticleTranslationRemovedEvent::class, 'method' => 'onArticleChanged'])
             ->tag('kernel.event_listener', ['event' => ArticleTranslationCopiedEvent::class, 'method' => 'onArticleChanged']);
 
-        $services->set('sulu_article.article_reindex_provider')
-            ->class('Sulu\Article\Infrastructure\Sulu\Search\ArticleReindexProvider')
+        $services->set('sulu_article.admin_article_reindex_provider')
+            ->class('Sulu\Article\Infrastructure\Sulu\Search\AdminArticleReindexProvider')
             ->args([
                 new Reference('doctrine.orm.entity_manager'),
             ])
