+++ packages/page/src/Infrastructure/Sulu/Search/AdminPageIndexListener.php
@@ -30,46 +30,23 @@ use Symfony\Component\Messenger\MessageBusInterface;
  * @internal this class is internal no backwards compatibility promise is given for this class
  *           use Symfony Dependency Injection to override or create your own Listener instead
  */
-final class PageIndexListener
+final class AdminPageIndexListener
 {
     public function __construct(
         private readonly MessageBusInterface $messageBus,
     ) {
     }
 
-    // Todo: Check if translation removed and translation restored event is needed.
     public function onPageChanged(PageCreatedEvent|PageModifiedEvent|PageRemovedEvent|PageRestoredEvent|PageTranslationRestoredEvent|PageTranslationAddedEvent|PageTranslationRemovedEvent $event): void
     {
-        $locale = $event->getResourceLocale();
-        $identifiers = [];
+        $resourceId = $event->getResourceId();
 
-        if ($event instanceof PageRemovedEvent) {
-            $locales = $event->getAllLocales();
+        $identifiers = \array_map(
+            fn (string $locale) => PageInterface::RESOURCE_KEY . '::' . $resourceId . '::' . $locale,
+            $this->getLocales($event),
+        );
 
-            if (!$locales) {
-                return;
-            }
-
-            foreach ($locales as $locale) {
-                $identifiers[] = PageInterface::RESOURCE_KEY . '::' . $event->getResourceId() . '::' . $locale;
-            }
-        } elseif ($event instanceof PageRestoredEvent) {
-            $page = $event->getPage();
-            $dimensionContentCollection = new DimensionContentCollection($page->getDimensionContents()->toArray(), [], PageDimensionContent::class);
-            $unlocalizedDimensionContent = $dimensionContentCollection->getDimensionContent(['locale' => null, 'stage' => 'draft']);
-
-            if (!$unlocalizedDimensionContent?->getAvailableLocales()) {
-                return;
-            }
-
-            foreach ($unlocalizedDimensionContent->getAvailableLocales() as $locale) {
-                $identifiers[] = PageInterface::RESOURCE_KEY . '::' . $event->getResourceId() . '::' . $locale;
-            }
-        } elseif ($locale) {
-            $identifiers[] = PageInterface::RESOURCE_KEY . '::' . $event->getResourceId() . '::' . $locale;
-        }
-
-        if (!$identifiers) {
+        if ([] === $identifiers) {
             return;
         }
 
@@ -78,5 +55,25 @@ final class PageIndexListener
                 ->withIndex('admin')
                 ->withIdentifiers($identifiers),
         );
+    }
+
+    /**
+     * @return string[]
+     */
+    private function getLocales(PageCreatedEvent|PageModifiedEvent|PageRemovedEvent|PageRestoredEvent|PageTranslationRestoredEvent|PageTranslationAddedEvent|PageTranslationRemovedEvent $event): array
+    {
+        if ($event instanceof PageRemovedEvent) {
+            return $event->getAllLocales() ?? [];
+        }
+
+        if ($event instanceof PageRestoredEvent) {
+            $page = $event->getPage();
+            $dimensionContentCollection = new DimensionContentCollection($page->getDimensionContents()->toArray(), [], PageDimensionContent::class);
+            $unlocalizedDimensionContent = $dimensionContentCollection->getDimensionContent(['locale' => null, 'stage' => 'draft']);
+
+            return $unlocalizedDimensionContent ? ($unlocalizedDimensionContent->getAvailableLocales() ?? []) : [];
+        }
+
+        return $event->getResourceLocale() ? [$event->getResourceLocale()] : [];
     }
 }
