+++ src/Sulu/Bundle/CategoryBundle/Resources/config/services.xml
@@ -80,8 +80,8 @@
         </service>
 
         <service
-                id="sulu_category.category_index_listener"
-                class="Sulu\Bundle\CategoryBundle\Infrastructure\Sulu\Search\CategoryIndexListener"
+                id="sulu_category.admin_category_index_listener"
+                class="Sulu\Bundle\CategoryBundle\Infrastructure\Sulu\Search\AdminCategoryIndexListener"
         >
             <argument type="service" id="sulu_message_bus"/>
             <tag name="kernel.event_listener" event="Sulu\Bundle\CategoryBundle\Domain\Event\CategoryCreatedEvent" method="onCategoryChanged"/>
@@ -92,8 +92,8 @@
         </service>
 
         <service
-                id="sulu_category.category_reindex_provider"
-                class="Sulu\Bundle\CategoryBundle\Infrastructure\Sulu\Search\CategoryReindexProvider"
+                id="sulu_category.admin_category_reindex_provider"
+                class="Sulu\Bundle\CategoryBundle\Infrastructure\Sulu\Search\AdminCategoryReindexProvider"
         >
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <tag name="cmsig_seal.reindex_provider"/>
