+++ src/Sulu/Bundle/ContactBundle/Infrastructure/Sulu/Search/AdminAccountReindexProvider.php
@@ -31,7 +31,7 @@ use Sulu\Bundle\ContactBundle\Entity\AccountInterface;
  * @internal this class is internal no backwards compatibility promise is given for this class
  *            use Symfony Dependency Injection to override or create your own ReindexProvider instead
  */
-final class AccountReindexProvider implements ReindexProviderInterface
+final class AdminAccountReindexProvider implements ReindexProviderInterface
 {
     /**
      * @var EntityRepository<AccountInterface>
