+++ src/Sulu/Bundle/ContactBundle/Infrastructure/Sulu/Search/AdminContactReindexProvider.php
@@ -32,7 +32,7 @@ use Sulu\Bundle\ContactBundle\Entity\ContactInterface;
  * @internal this class is internal no backwards compatibility promise is given for this class
  *            use Symfony Dependency Injection to override or create your own ReindexProvider instead
  */
-final class ContactReindexProvider implements ReindexProviderInterface
+final class AdminContactReindexProvider implements ReindexProviderInterface
 {
     /**
      * @var EntityRepository<ContactInterface>
