+++ src/Sulu/Bundle/ContactBundle/Resources/config/services.xml
@@ -212,8 +212,8 @@
         </service>
 
         <service
-                id="sulu_contact.contact_index_listener"
-                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\ContactIndexListener"
+                id="sulu_contact.admin_contact_index_listener"
+                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AdminContactIndexListener"
         >
             <argument type="service" id="sulu_message_bus"/>
             <tag name="kernel.event_listener" event="Sulu\Bundle\ContactBundle\Domain\Event\AccountCreatedEvent" method="onAccountChanged"/>
@@ -227,16 +227,16 @@
         </service>
 
         <service
-                id="sulu_contact.account_reindex_provider"
-                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AccountReindexProvider"
+                id="sulu_contact.admin_account_reindex_provider"
+                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AdminAccountReindexProvider"
         >
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <tag name="cmsig_seal.reindex_provider"/>
         </service>
 
         <service
-                id="sulu_contact.contact_reindex_provider"
-                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\ContactReindexProvider"
+                id="sulu_contact.admin_contact_reindex_provider"
+                class="Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AdminContactReindexProvider"
         >
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <tag name="cmsig_seal.reindex_provider"/>
