+++ src/Sulu/Bundle/ContactBundle/Tests/Functional/Infrastructure/Sulu/Search/AdminContactReindexProviderTest.php
@@ -15,21 +15,24 @@ namespace Sulu\Bundle\ContactBundle\Tests\Functional\Infrastructure\Sulu\Search;
 
 use CmsIg\Seal\Reindex\ReindexConfig;
 use Doctrine\ORM\EntityManagerInterface;
-use Sulu\Bundle\ContactBundle\Entity\Account;
-use Sulu\Bundle\ContactBundle\Entity\AccountInterface;
-use Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AccountReindexProvider;
+use Sulu\Bundle\ContactBundle\Entity\Contact;
+use Sulu\Bundle\ContactBundle\Entity\ContactInterface;
+use Sulu\Bundle\ContactBundle\Infrastructure\Sulu\Search\AdminContactReindexProvider;
 use Sulu\Bundle\MediaBundle\Entity\Collection;
 use Sulu\Bundle\MediaBundle\Entity\CollectionType;
 use Sulu\Bundle\MediaBundle\Entity\File;
 use Sulu\Bundle\MediaBundle\Entity\FileVersion;
 use Sulu\Bundle\MediaBundle\Entity\Media;
 use Sulu\Bundle\MediaBundle\Entity\MediaType;
+use Sulu\Bundle\TestBundle\Testing\SetGetPrivatePropertyTrait;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
 
-class AccountReindexProviderTest extends SuluTestCase
+class AdminContactReindexProviderTest extends SuluTestCase
 {
+    use SetGetPrivatePropertyTrait;
+
     private EntityManagerInterface $entityManager;
-    private AccountReindexProvider $provider;
+    private AdminContactReindexProvider $provider;
 
     private MediaType $imageType;
 
@@ -38,7 +41,7 @@ class AccountReindexProviderTest extends SuluTestCase
     protected function setUp(): void
     {
         $this->entityManager = $this->getEntityManager();
-        $this->provider = new AccountReindexProvider($this->entityManager);
+        $this->provider = new AdminContactReindexProvider($this->entityManager);
         $this->purgeDatabase();
 
         $this->imageType = new MediaType();
@@ -61,13 +64,13 @@ class AccountReindexProviderTest extends SuluTestCase
 
     public function testGetIndex(): void
     {
-        $this->assertSame('admin', AccountReindexProvider::getIndex());
+        $this->assertSame('admin', AdminContactReindexProvider::getIndex());
     }
 
     public function testTotal(): void
     {
-        $this->createAccount('Count 1');
-        $this->createAccount('Count 2');
+        $this->createContact();
+        $this->createContact();
 
         $this->entityManager->flush();
 
@@ -76,8 +79,8 @@ class AccountReindexProviderTest extends SuluTestCase
 
     public function testProvideAll(): void
     {
-        $account1 = $this->createAccount('Test Account 1', 'Media 1');
-        $account2 = $this->createAccount('Test Account 2');
+        $contact1 = $this->createContact('Tom', 'Turbo', 'avatar1');
+        $contact2 = $this->createContact();
 
         $this->entityManager->flush();
 
@@ -85,16 +88,16 @@ class AccountReindexProviderTest extends SuluTestCase
         $changedDateString2 = '2024-06-01 15:30:00';
 
         $connection = self::getEntityManager()->getConnection();
-        $sql = 'UPDATE co_accounts SET changed = :changed WHERE id = :id';
+        $sql = 'UPDATE co_contacts SET changed = :changed WHERE id = :id';
 
         $connection->executeStatement($sql, [
             'changed' => $changedDateString1,
-            'id' => $account1->getId(),
+            'id' => $contact1->getId(),
         ]);
 
         $connection->executeStatement($sql, [
             'changed' => $changedDateString2,
-            'id' => $account2->getId(),
+            'id' => $contact2->getId(),
         ]);
 
         $config = ReindexConfig::create()->withIndex('admin');
@@ -105,22 +108,22 @@ class AccountReindexProviderTest extends SuluTestCase
         $this->assertSame(
             [
                 [
-                    'id' => AccountInterface::RESOURCE_KEY . '::' . $account1->getId(),
-                    'resourceKey' => AccountInterface::RESOURCE_KEY,
-                    'resourceId' => (string) $account1->getId(),
-                    'mediaId' => (string) $account1->getLogo()?->getId(),
+                    'id' => ContactInterface::RESOURCE_KEY . '::' . $contact1->getId(),
+                    'resourceKey' => ContactInterface::RESOURCE_KEY,
+                    'resourceId' => (string) $contact1->getId(),
+                    'mediaId' => (string) $contact1->getAvatar()?->getId(),
                     'changedAt' => (new \DateTimeImmutable($changedDateString1))->format('c'),
                     'createdAt' => (new \DateTimeImmutable('2000-01-01 12:00:00'))->format('c'),
-                    'title' => $account1->getName(),
+                    'title' => $contact1->getFullName(),
                 ],
                 [
-                    'id' => AccountInterface::RESOURCE_KEY . '::' . $account2->getId(),
-                    'resourceKey' => AccountInterface::RESOURCE_KEY,
-                    'resourceId' => (string) $account2->getId(),
+                    'id' => ContactInterface::RESOURCE_KEY . '::' . $contact2->getId(),
+                    'resourceKey' => ContactInterface::RESOURCE_KEY,
+                    'resourceId' => (string) $contact2->getId(),
                     'mediaId' => '',
                     'changedAt' => (new \DateTimeImmutable($changedDateString2))->format('c'),
                     'createdAt' => (new \DateTimeImmutable('2000-01-01 12:00:00'))->format('c'),
-                    'title' => $account2->getName(),
+                    'title' => $contact2->getFullName(),
                 ],
             ],
             [...$results],
@@ -129,15 +132,15 @@ class AccountReindexProviderTest extends SuluTestCase
 
     public function testProvideWithSpecificIdentifiers(): void
     {
-        $account1 = $this->createAccount('Account One');
-        $account2 = $this->createAccount('Account Two');
-        $account3 = $this->createAccount('Account Three');
+        $contact1 = $this->createContact();
+        $contact2 = $this->createContact('Fritz', 'Fantom');
+        $contact3 = $this->createContact('Thomas', 'Brezina');
 
         $this->entityManager->flush();
 
         $identifiers = [
-            AccountInterface::RESOURCE_KEY . '::' . $account1->getId(),
-            AccountInterface::RESOURCE_KEY . '::' . $account3->getId(),
+            ContactInterface::RESOURCE_KEY . '::' . $contact1->getId(),
+            ContactInterface::RESOURCE_KEY . '::' . $contact3->getId(),
         ];
 
         $config = ReindexConfig::create()
@@ -149,25 +152,26 @@ class AccountReindexProviderTest extends SuluTestCase
         $this->assertCount(2, $results);
 
         $resultTitles = \array_column($results, 'title');
-        $this->assertContains('Account One', $resultTitles);
-        $this->assertContains('Account Three', $resultTitles);
-        $this->assertNotContains('Account Two', $resultTitles);
+        $this->assertContains('Tom Turbo', $resultTitles);
+        $this->assertContains('Thomas Brezina', $resultTitles);
+        $this->assertNotContains('Fritz Fantom', $resultTitles);
     }
 
-    private function createAccount(string $name, ?string $mediaName = null): Account
+    private function createContact(string $firstName = 'Tom', string $lastName = 'Turbo', ?string $avatarName = null): Contact
     {
-        $account = new Account();
-        $account->setName($name);
-        $account->setCreated(new \DateTimeImmutable('2000-01-01 12:00:00'));
+        $contact = new Contact();
+        $contact->setFirstName($firstName);
+        $contact->setLastName($lastName);
+        $contact->setCreated(new \DateTimeImmutable('2000-01-01 12:00:00'));
 
-        if ($mediaName) {
-            $media = $this->createMedia($mediaName);
-            $account->setLogo($media);
+        if (null !== $avatarName) {
+            $media = $this->createMedia($avatarName);
+            $contact->setAvatar($media);
         }
 
-        $this->entityManager->persist($account);
+        $this->entityManager->persist($contact);
 
-        return $account;
+        return $contact;
     }
 
     private function createMedia(string $name): Media
